requirements.txt
pandas 
matplotlib
notebook
python-box==6.0.2
ensure
seaborn
plotly
cassandra-driver
-e .